adding checks for product id

+++ backend/controllers/product.controller.js
@@ -36,7 +36,7 @@ export const updateProduct = async (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res
       .status(404)
-      .json({ success: false, message: 'product not found' });
+      .json({ success: false, message: 'Invalid product id' });
   }
 
   try {
@@ -52,11 +52,16 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
   console.log(id);
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: 'Invalid product id' });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: 'Product deleted' });
   } catch (error) {
     console.log(`error in deleting product: ${error.message}`);
-    res.status(404).json({ success: false, message: 'Product not found' });
+    res.status(500).json({ success: false, message: 'Server error' });
   }
 };
